+++ ceammc/CHANGELOG.md
@@ -22,6 +22,7 @@
   - [at(, [front(, [back( methods added to data.list/data.mlist
   - [choose( method added to get random element from \[data.set\], \[local.set\] and \[global.set\]
   - [choose( method added to get random element from \[data.dict\], \[local.dict\] and \[global.dict\]
+  - [pause( method added to \[array.grainer~\]
 - misc:
   - conv.lin2exp: list support added
   - conv.lin2curve: list support added